Adjust test to constrain the evaluated conditions
As per the reasoning recorded in
https://github.com/datalad/datalad/issues/1811#issuecomment-1114298954
the test ensuring the continue fix of #1811 is actually
verifying a situation that is not fully desirable. It prevents
pushing of thew 'git-annex' branch whenever a repo is on a detached
HEAD.

This change let's the test run on a plain Git repo, where there is
indeed nothing to push in this case.