ENH: Enable progress bars for unlock, fixes #6691
An unlocking operation can take a considerable amount of
time, but doesn't provide feedback at all (see #6691).
By switching from _call_annex_records(), which processes
the full bunch of files at once, to _call_annex_records_items(),
which yields as soon as results are available, we can
construct a progress bar with a file counter around the
unlock operation.

The progress reporting is slightly complicated by the
fact that git annex' --json records for each file are
output before a final 'recording state in git' operations,
which can, if a large number of files was unlocked, add
considerably to the execution time of the command after
all files have been unlocked already and the progress bar
would be at 100%. Thus, this change adds an additional
progress update when all files are processed to report
on the final recording step.