MNT: Factor out helper function used in core and next from ria functionality
create-sibling-ria had in-line code to check for datasets with matching sibling names
that datalad-next refactored into a helper function.
This PR adds the helper function created in datalad-next to datalad/distribution/utils
and uses it instead of the previous in-line code to make the code more modular.
A companion PR in datalad-next may import the helper from datalad core.