Lazy `datalad.ui.ui` instantiation
So far, UI setup code was triggered immediately and unconditionally on
`import datalad`. This made it hard to make it configurable, because
the decision making is running before any of the basic configuration
layers are up (see https://github.com/datalad/datalad/issues/6714 for
some facet of this).

This change refactors the `_UI_Switcher` class to delay any decision
making to the point where client code actually requests UI
functionality.

This change in its own is not sufficient to effectively delay UI
instantiation (again see
https://github.com/datalad/datalad/issues/6714), but sets the stage for
subsequent improvements.