Change DATALAD_LOG_TRACEBACK behavior
This change aims to make the output more compact and more
informative.

The stack frame of the log handler re-emitting a message is not longer
included in the report. This is a uniform internal behavior that is
identical for all messages, and not useful for debugging with this
utility (because the previously established removal of internal
logging handling from the traceback already made it unsuitable for
debugging logging code -- which is perfectly fine).

Was
```
% DATALAD_LOG_TRACEBACK=collide dl -l debug status
[DEBUG  ] datalad:33>main:86>parser:119,369>log:223  Command line args 1st pass for DataLad 0.16.2+190.gd657c17fe. Parsed: Namespace() Unparsed: ['status']
[DEBUG  ] ...>main:94>log:223  Parsing known args among ['/home/mih/env/datalad-dev/bin/datalad', '-l', 'debug', 'status']
```

Now
```
% DATALAD_LOG_TRACEBACK=collide dl -l debug status
[DEBUG  ] datalad:33>main:86>parser:119,369  Command line args 1st pass for DataLad 0.16.2+194.g83e04f37f.dirty. Parsed: Namespace() Unparsed: ['status']
[DEBUG  ] …>main:94  Parsing known args among ['/home/mih/env/datalad-dev/bin/datalad', '-l', 'debug', 'status']
```

`DATALAD_LOG_TRACEBACK=1` will now typically report the origin of
any log message, rather than a location in `log.py`, a more natural
behavior IMHO. Due to the previously established removal of "common"
stack frames, the traceback depth already could not be interpreted
in a strict manner.

Was

```
% DATALAD_LOG_TRACEBACK=1 dl -l debug status
[DEBUG  ] ...>log:223  Command line args 1st pass for DataLad 0.16.2+190.gd657c17fe. Parsed: Namespace() Unparsed: ['status']
[DEBUG  ] ...>log:223  Parsing known args among ['/home/mih/env/datalad-dev/bin/datalad', '-l', 'debug', 'status']
```

Now

```
% DATALAD_LOG_TRACEBACK=1 dl -l debug status
[DEBUG  ] parser:369  Command line args 1st pass for DataLad 0.16.2+194.g83e04f37f.dirty. Parsed: Namespace() Unparsed: ['status']
[DEBUG  ] main:94  Parsing known args among ['/home/mih/env/datalad-dev/bin/datalad', '-l', 'debug', 'status']
```

The actual behavior of this switch, and that it is not a proper
config switch, is now documented.

Closes datalad/datalad#6744