keep line endings intact as long as possible
This commit extends LineSplitter with the ability
to keep line-endings in the resulting lines.

The new LineSplitter is used in
GitRepo._generator_call_git() and improves
GitRepo.call_git_items_(). This allows to use
call_git() and call_git_items_() exchangable. That
means the output of call_git()[0] and
"".join(call_git_items_(..., keep_ends=True)) are
identical.