Speed up saving of removals and improve reporting
The previous implementation in `GitRepo._save()` used `git-rm` to
handle the staging of removed files. However, this porcelain command
also handles the actual removal of files from the working tree.
This is unnecessary here, because only items are processed that
have already found to be removed from the working tree. Therefore,
the more lightweight `update-index` is used now, and provides measurable
speed-ups. I found between 5-10% faster `datalad-save` runtimes when
removing 2k file entries from a dataset on a lightning-fast NVME drive.
Because `GitRepo` methods using the `normalize_paths` decorator are
avoided not, the speed-up should be even more substantial on slower
file systems.

This change also improves the reporting of `datalad-save` when saving
the removal of a subdataset. Previously, it was reported as

```
% datalad save -d demo
delete(ok): sub (file)
```

with the confusing `file` type property. Now it is properly reported as
a dataset:

```
% datalad save -d demo
delete(ok): sub (dataset)
```

Moreover, `save` now also reports saving the deletion of files, when
the deletion was already staged.