BF: use --python-match minor option in new datalad-installer release to match outside version of python
@bpoldrack spotted that we are not actually testing against the python version
we set in our matrix runs of travis CI setup.  with datalad-installer 0.9.0 we
can now match to minor (or try even more detailed match) portion of the version
of system wide python. Again -- it would not be system wide python intallation,
but conda installation of python matching in minor portion of the version, which
is what we care in our matrix