Make sure a subdataset is saved with a complete .gitmodules record
Previously, the test for adding a gitmodules record was incomplete
and relied on the assumption that for a new subdataset the type
`dataset` was not determined by the initial `status()` inspection.
However, this is not the case, when an explicit subdataset path
with a trailing path-separator is given to `save()`.

This change amends the test condition to also test for the state
being `untracked`. In principle this alone should be a sufficient
condition. However, due to the complexity of the decision making
in the present implementation, the previous condition as also
kept as an alternative, in order to prevent unexpected side-effect.

Fixes datalad/datalad#6547