Improved support for saving typechanges
This changeset brings a tests for verifying that `save()` can handle the
following type changes in a dataset:

- replace a file with a directory (that has content)
- replace a file with a (sub)dataset
- replace a directory with a (sub)dataset
- replace a subdataset with a file

`GitRepo.save_()` was fixed to detect and handle replacing a file
with a (sub)dataset.

At the moment the (rather complex case of) replacing a directory that
contains a registered subdataset with a new subdataset requires two
calls to `save()`, instead of being able bring the dataset to a clean
state in one go. This first call automatically unregisters the vanished
subdataset, and the second call registers the new one.

Fixes datalad/datalad#5418