Enable `status/diff` to detect type-changes
Previously, type-changes (when a nature of a repository component
changes in the worktree, e.g., a subdataset is suddenly a file) where
not detected by `status/diff()`, because the `gitsha` report in the
does not differ between the "from" and "to" state.

This change enables this detection by combination this particular
characteristic with a "modified in working tree" indicator to
identify typechanges.

In such cases the previous `type` is not reported as the `prev_type`
property. The new type is determined by a dedicated call to `git
diff-files`, one per each typechange -- under the assumption that such
dataset modifications are rare.

For such cases, the `gitsha` property is no longer included, because
it is not provided by Git and only expensive to obtain. Previously,
the `gitsha` of the previous object of the old type was wrongly
reported.

Fixes datalad/datalad#6791