explain name selection in comment
This commit adds a comment to explain why we change the
name of the destination file, when copying a file with
obscure filename via ssh.

The reason is that the test setup maps the source and
destination directories to the same directory on the
test machine, i.e. the file is copied onto itself. With
ssh version 9, this leads to an empty file.