DOC: fix and move docstring reference into design docs
This change moves information on how docstrings are
to be designed from CONTRIBUTING to the Design docs
as the Design docs have become a place to consolidate
information like this and already contain a section
on docstrings. At the same time, it fixes the link
to the numpy docstring standard that did not resolve
anymore, and replaces the 'Documentation' section in
CONTRIBUTING, which only contained docstring info, with
a more general pointer to the design documents further
up in the file.
Fixes #6801