quote ssh filenames depending on ssh version
This commit adds a very rough ssh-version detection
to deal with different filename quoting requirements
in ssh version <= 8 and in ssh version 9.

This fixes the ssh copy tests error that is
described in issue #6806.

This is a rough first draft to verify the overall
viability and to test the waters in CI. The version
parsing will have to be improved.