BF: provide /dev/null as --git-dir instead of no value
As it was reported in https://github.com/datalad/datalad/issues/6815
and inquired from git folks for a better solution (no reply yet) at
https://lore.kernel.org/git/YscCKuoDPBbs4iPX@lena.dartmouth.edu/T/#u
whenever we provide an empty value for --git-dir, iff there is includeIf
statement in config, then git starts to demand a non-empty value for that
setting.  Unfortunately RTFM does not give any clear answer either that is
a legit or not. So I have decided to replace one questionable use with
another which does not cause an error.

Anther alternative I found we could have used two separate invocations of
--system and --global, but we cannot use them at the same time.