XFAIL test_runner_parametrized_protocol on python3.8 when getting duplicate output
This should close (unresolved) but not fix https://github.com/datalad/datalad/issues/4921
which seems to happen so far rarely on conda on 3.8 python only. If some brave soul
manages to figure it out -- great. But I would not break our knuckles to see it fixed
since seems to be very particular.  I made xfail also very particular (not over entire
test, only for 3.8 and only for the 55 output) so we do not miss any other problem