Align `status/diff` reporting behavior with `git-file-files`
Git does not report the previously recorded gitsha for content that
is modified in the worktree. This change removes the `gitshasum`
report from such a result record, in order to match this behavior.

With a worktree modification for a path, reporting the old `gitshasum`
will often just be wrong (except for type changes).

Fixes datalad/datalad#6875