RF: remove splitting remotes by space since was never supported before
@bpoldrack shines the archaeological light on the situation. We used to have
get return all values by default, that caused issues, so get_all=False was
introduced and thus broke prior behavior here which was to have all of the
siblings pushed to (I guess the test didn't test for multiple). So I will push
here a commit which replaces all the above with simple addition of
get_all=True.