RF: keep warning if really working with a URL, add "guessing" option to _pr_to_fields
That "guessing" would add semantic on the purpose of the invocation
of that function, and if we are guessing only, there is no need
to scarry user with the warning.  But if it is invoked in the other
use case where it is to parse URL (e.g. having decided it is a URL),
then it would still issue warning