BF: do not pretend we know what repo version git-annex would upgrade to
Since 10.20220724-5-g63cef2ae0 git-annex started to upgrade to v10 even though
v8 is still supported.  Our hard-coded mapping we relied on to report to the
user that upgrade is too happen would not allow for neede flexibility so
I removed it and ajusted INFO message to report what we know -- that upgrade
would be to at least the first supported version.  Test also was adjusted accordingly
while still testing prior behavior