RF: refactor test_extensions to centralize ext specifics, use extension"s own tox.ini
Original form started to accumulate heavy "if" statements and hard to
follow logic because of them.  Also now we need to refactor to install
not from master.zip but to have a clone of the repository so we can
get access to extension"s tox.ini.

In proposed form I use bash case/esac since it provides best flexibility
IMHO to define extension specific settings to be reused in the next
steps.