ENH: warn once if filesystem detail could not be determined
Even if it would be too noisy on windows -- would still be better than
just silently not working correctly (as it was due to a typo).

And if psutil is not installed, we would get it logged only at debug level. But
otherwise, if something abnormal -- at warning level and reported once
(although could be different for different filesystems, but at least would
alert to a possible next typo)