ENH: GitRepo -- make use of --pathspec-from-file where possible and chunking needed
Chunking is a kludge to workaround the problem of DataLad needing to provide a
long list of files to git/git-annex commands.  Whenever git-annex has --batch mode
which we use typically for commands which could get lots of file paths arguments, we
have not used any special means like that for git commands.  Many commands support
--pathspec-file-nul and --pathspec-from-file=FILE options to pass a list of pathspec(s)
to operate on via a FILE instead of command line.

In this PR we decide to pass via such a file in cases where more than a single
chunk is necessary.