RF: get_modules_ - avoid O(N^2) if paths provided via use of get_parent_paths
that function was created for get_content_info for that specific reason
(performance). Prior commits avoided a hypothetical infinite recursion (my code
analysis says it was not possible, see
https://github.com/datalad/datalad/issues/6941#issuecomment-1222385283) and
performance issues by introducing similar logic directly in the code of
get_modules_.   This commit RFs that back to use get_parent_paths and adds
rudimentary tests for invocation with paths limiting.  But tests showed that
such approach "as is" is buggy.  It is primarily due to the incorrect handling
for whenever "paths" is provided (not None) but empty list, e.g. due to
filtering of the values etc.  Subsequent commit will deal with that issue.