BF: make logic more consistent for files=[] argument (which is False but not None)
Similar problem was resolved previously at higher level in
http://github.com/datalad/datalad/commit/ec0243c92822f36ada5e87557eb9f5f53929c9ff
for get_content_info.  But IMHO it could and should be resolved at the
lowest/common point of invocation, that is why I added it to
_generator_call_git which is underlying most (if not all, hard to tell now ;) )
git invocations.

I also fixed a few other invocations where it was pure `if files` or `if paths`
instead of correctly checking against None.