report missing gen4-metadata correctly
This commit fixes the "impossible" result records
that are returned if no gen4-metadata is available,
but a gen4-metadata is querried.

Before this commit, all queried types would be
included in a tuple in a single result record.
Now a result record is created for every requested
type, i.e. "file" or "dataset".