Set SOURCE_DATE_EPOCH and DATALAD_SOURCE_EPOCH explicitly
As build showed, prior approach resulted in no fixed date. May be it is because
we did not include any dpkg include files. I decided not to wonder and just
make it explicit through parsing changelog. And that is why I am not changing
changelog and date so we could test that build did it right