Remove RIA functionality
This changeset is part of an effort to strip datalad-core off
non-essential components in order to reduce technical debt and
improve maintainability.

RIA (the concept of a dataset store, the git-annex special remote
the URL handling) is not an essential component, but an application
built on the datalad core.

https://github.com/datalad/datalad/pull/7017 shows how it can be
completely separated from the core, because it is only integrated
in within the `clone()` command -- all other code is located in
pieces that can be installed in parallel.

At the same time RIA is large (~6k lines of code), and has accumulated
as significant number of unaddressed technical and conceptual issues.
The amount to more than 7% of the total number of currently open issues
in datalad-core. Moreoever, it is severly underdocumented.

This change documents the code pieces that make up RIA, and the impact
of their removal from core. They would move into an extension package,
and remain available to users.

The most significant impact of this removal is the removal of a set
of tests of core functionality that made use of RIA in one form or
another. This loss could trun out to be substantial and require
compensatory actions. However, these would have been needed anyways
(and or simply harder to ignore now), because the RIA tooling did not
work across all target platforms, and the majority of these tests
were skipped on one or more platform.