Let `create` touch the dataset root, if not saving in parent dataset
This communicates the "nature change" of the directory (from plain to
dataset) to systems like inotify.

Without this, a dataset creation in an existing, empty subdirectory
would go undiscovered, unless the empty subdirectory is explicitly
watched itself.

With this change, that comes at virtually no additional cost, it is
sufficient to watch the parent directory if the previously empty
target subdirectory.