Give more information about used token and performed actions
We have an odd behavior in being unable to comment on issues, and looks
like if GITHUB_TOKEN was set to an empty value. Here we will error out if
empty and otherwise reveal the beginning of it which by itself should not
be leaking too much of a secret to do any harm.

Also print about actions being performed since now it is just silently
does its deeds

[ci skip]
[appveyor skip]