ENH: DATALAD_TESTS_GITCONFIG to provide custom settings for git during tests
Although could be just once in a lifetime bug, we just got into an odd
issue with an unlocked file with content being marked as modified
https://git-annex.branchable.com/bugs/reports_file___34__modified__34___whenever_it_is_not/#comment-4e61c4c0f493e98f9416e09de4cfb33f
which seems to boil down to have a rarely used `annex.stalldetection` set.

By adding one more env variable for TESTS we can now add such specific
setting to be tested in the tests.