improve git-annex failure handling
This commit adds code to handle the situation
where git-annex in batch-mode does not signal
an error, here: an unknown remote, by returning
a JSON object with the error description, but
by exiting with an error code.

This has become necessary after `BatchedCommand`
was modified so that it reports errors of
subprocesses to its callers.

Remark:
The code added in this commit only fixes
`AnnexRepo.is_available()`. It maps all error
exits of git-annex to "non-availability", which
might be too unspecific. There also might be other
places in the code, where process-exits should
be caught and mapped onto method-results.