BF: don't write output from within sshrun
`datalad sshrun` explicitly calls SSH with `log_output=False` which
results in the use of `NoCapture` protocol with the runner. Meaning,
stdout/stderr of SSH is written out anyway already. When SSH returns,
`sshrun` tried to write both to its stdout/stderr. But: It could not
possibly have anything to write. That would not be an issue in and of
itself, but `sshrun` is not necessarily used directly. In particular it
is called by `git` (due to `GIT_SSH_COMMAND=datalad sshrun`). This
resulted in a problem when apparently `git` has closed the pipe to its
ssh executable (`sshrun`) already and we tried to write to it (although
we really didn't even have something to write).

This ultimately led to issue #6599, where the actual `ssh ...
git-upload-pack` execution succeeded and returned 0, but `datalad
sshrun` itself produced a broken pipe error trying to write to stdout
and hence returning non-zero.

It's not entirely clear when exactly this happens. It may be depend on
git version when the pipe is closed as the failing builds are running
2.35.1 (MacOS on appveyor) whereas otherbuilds have either newer or
older versions of git. In any case: There can't be anything to write out
to begin with, so don't even try.

Also: Make it clear in the code, that and why we don't expect any
captured output from the SSH subprocess by not storing the empty return
value, so future changes (and debuggers) don't falsely assume that
1. Output can simply be captured (with existing protocols) or
2. The returned value would currently be of any use simply b/c it's
   there.

(Closes #6599)
(Closes #7078)