Fix test helpers for results
The helpers for testing result records had a couple of bugs.
Most noteably, some (e.g. `assert_status`) would pass when no result was
given at all. Therefore command calls that did not return any result
were passing assertions about the `status` field of their results.

Furthermore, most of those helpers did not properly deal with all three
"representations" of results that are (and need to be) passed to them:
1. list of dict
2. single dict
3. generator

That's because `ensure_list` is used with the incoming results, but this
turns dicts that are not encapsulated by a list into a list of their
keys rather than a single-item list of dict. Reason for that is, that a
dict is iterable.

This patch should fix this.