Report repository name in dry run
This adds an alternative result message in `create_sibling_ghlike`
when used with `dry_run`. The mesage now reports the name of the
repository which would be created (useful when running
recursively). Previously, the message would say "sibling repository
<siblingname> created at None" (because dry run makes no requests,
repository url was not known).

Maybe it would be more logical to add this message to the result
returned by `repo_create_request` earlier in the command chain (which,
for dry run, only returns datails of the request it *would* make). But
this function doesn't know the sibling name, and adding it as a
parameter didn't feel good either.