improve resilience of stalling test.
This commit extends the time after which we believe
to see a stalling thread to at least 5 seconds. This
time will only be exhausted, if the thread is stalling,
the duration of a passing test is usually in the range
of a few dozen milliseconds.

The commit adds output processing in order to actually
start of few runner-related threads and perform
thread-based asynchronous data processing.

The huge time limit will hopefully reduce the number
of false positives greatly, though is cannot *guarantee*
the absence of false positives, it makes them unlikely,
i.e. less than 1 false positive per 10e5 executions.