BF: Fix reported path to ria-layout-version in execption
Probing for the existence of a ria_layout_version file internally
constructs a target_ri variable only used for reporting the path
to the file. We noticed recently that this reporting was off,
duplicating paths, e.g.:

[INFO   ] RIA store unavailable. -caused by- file:///tmp/myriastore/tmp/myriastore/ria-layout-version not found. -caused by- [Errno 2] No such file or directory: '/tmp/myriastore/ria-layout-version'

This duplication should be fixed with this small change. It
extracts the file name (ria_layout_version) and appends it to the ria
store location, instead of appending the entire path to it.